receiver-requirements.txt
numpy
pyaudio
matplotlib
gpsd-py3
folium
